requirements.txt
nextcord==2.0.0a3
pickleDB==0.9.2
PyYAML==5.3.1